MB-68266: Fix narrowing conversion warnings in tests
Replace implicit narrowing conversions with explicit casts using
gsl::narrow or appropriate type changes:

- Change loop counters to uint8_t where matching function params
 - Use std::size_t for count return values and loop variables
 - Add gsl::narrow for test values known to fit in smaller types
 - Add recvDcpBufferAck(size_t) overload to avoid caller warnings

Also fix logic bug in snapshot_tests where loop counter was being
reassigned with arithmetic() return value instead of incrementing.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/244298
Tested-by: Build Bot <build@couchbase.com>

+++ tests/mcbp/mcbp_dcp_test.cc
@@ -894,7 +894,7 @@ TEST_P(DcpSetVbucketStateValidatorTest, CorrectMessage) {
 TEST_P(DcpSetVbucketStateValidatorTest, LegalValues) {
     cb::mcbp::RequestBuilder builder({blob, sizeof(blob)}, true);
 
-    for (int ii = 1; ii < 5; ++ii) {
+    for (uint8_t ii = 1; ii < 5; ++ii) {
         cb::mcbp::request::DcpSetVBucketState extras;
         extras.setState(ii);
         builder.setExtras(extras.getBuffer());
